Fix payload fairing thickness calculation
When the middle of the payload fairing is wider than the base, fairing
thickness was incorrectly calculated. This allowed players to
effectively cheese the thickness by making the base arbitrarily small,
allowing for arbitrarily thin fairings.

Now, the fairing thickness should be based on the maximum width of the
fairing instead of the base.

A little bit of refactoring was required to make this work, since
we need the scan data earlier to determine the correct payload fairing
width.